requirements.txt
aniso8601==9.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
deprecation==2.1.0
docopt==0.6.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Executor==0.9.4
Flask-RESTful==0.3.8
Flask-SSE==1.0.0
future==0.17.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
num2words==0.5.10
numpy==1.19.5
packaging==20.9
pyaml==19.12.0
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.22.2.post1
scipy==1.5.4
six==1.15.0
sklearn-crfsuite==0.3.6
gunicorn
snips-nlu==0.20.2
snips-nlu-utils==0.9.1
tabulate==0.8.9
tqdm==4.59.0
urllib3==1.26.4
Werkzeug==1.0.1
